+++ crates/qbz-app/Cargo.toml
@@ -24,3 +24,6 @@ chrono = { workspace = true }
 md-5 = { workspace = true }
 base64 = { workspace = true }
 rand = "0.10"
+
+# TLS provider selection
+rustls = { version = "0.23", features = ["aws_lc_rs"] }
